src/server/requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.12.0
rsa==4.0
six==1.14.0
Telethon==1.11.3
Werkzeug==1.0.0
